+++ example/src/main/java/com/muellerwulff/viewmodelextensions/example/MainActivity.kt
@@ -4,10 +4,8 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
+import androidx.lifecycle.observe
 import com.muellerwulff.viewmodelextensions.example.databinding.ActivityMainBinding
-import com.muellerwulff.viewmodelextensions.observe
-import com.muellerwulff.viewmodelextensions.observeRequired
-import com.muellerwulff.viewmodelextensions.requireValue
 import com.muellerwulff.viewmodelextensions.viewModelsCustom
 
 class MainActivity : AppCompatActivity() {
@@ -20,10 +18,10 @@ class MainActivity : AppCompatActivity() {
         binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
         binding.version = model.version
 
-        model.number.observeRequired(this) {
+        model.number.observe(this) {
             binding.number = it.toString()
         }
-        model.numberMultitude.observeRequired(this) {
+        model.numberMultitude.observe(this) {
             binding.multitude = it.toString()
         }
         model.warning.observe(this) { event ->
@@ -37,6 +35,6 @@ class MainActivity : AppCompatActivity() {
 
     override fun onResume() {
         super.onResume()
-        binding.number = model.number.requireValue().toString()
+        binding.number = model.number.value!!.toString()
     }
 }